Support inline xpack declaration in direct_html
Sometimes we declare the xpack tag inline like so:
```
[float]
=== [xpack]#Foo#
```

This renders the xpack tag in those cases.
