+++ integtest/html_diff
@@ -179,6 +179,12 @@ def normalize_html(html):
         parent = e.parent
         e.extract()
         parent.smooth()
+    # Docbook sometimes renders the xpack tag *after* the edit me link but
+    # asciidoctor always renders it before the edit me link. Either way is fine
+    # so we ignore the difference.
+    for e in soup.select("a.edit_me"):
+        parent = e.parent
+        parent.append(e.extract())
 
     # Remove empty "class" attributes and sort the listed classes.
     for e in soup.select('*'):
