+++ resources/asciidoctor/lib/docbook_compat/extension.rb
@@ -4,8 +4,10 @@ require 'asciidoctor/extensions'
 require_relative '../delegating_converter'
 require_relative '../strip_tags'
 require_relative 'convert_admonition'
-require_relative 'convert_document'
 require_relative 'convert_dlist'
+require_relative 'convert_document'
+require_relative 'convert_floating_title'
+require_relative 'convert_inline_quoted'
 require_relative 'convert_links'
 require_relative 'convert_listing'
 require_relative 'convert_lists'
@@ -31,8 +33,10 @@ module DocbookCompat
   # A Converter implementation that emulates Elastic's docbook generated html.
   class Converter < DelegatingConverter
     include ConvertAdmonition
-    include ConvertDocument
     include ConvertDList
+    include ConvertDocument
+    include ConvertFloatingTitle
+    include ConvertInlineQuoted
     include ConvertLinks
     include ConvertListing
     include ConvertLists
@@ -53,31 +57,6 @@ module DocbookCompat
         #{node.content}
         </div>
       HTML
-    end
-
-    def convert_floating_title(node)
-      tag_name = %(h#{node.level + 1})
-      # Asciidoctor's standard is to put the id on the header tag but docbook
-      # puts it in its own anchor tag.
-      anchor = node.id ? %(<a id="#{node.id}"></a>) : ''
-      classes = [node.role].compact
-      classes_html = classes.empty? ? '' : " class=#{classes.join ' '}"
-      <<~HTML
-        <#{tag_name}#{classes_html}>#{anchor}#{node.title}#{node.attr 'edit_me_link', ''}#{xpack_tag node}</#{tag_name}>
-      HTML
-    end
-
-    def convert_inline_quoted(node)
-      case node.type
-      when :monospaced
-        node.attributes['role'] ||= 'literal'
-        yield
-      when :strong
-        # Docbook's "strong" rendering is comically repetitive.....
-        %(<span class="strong strong"><strong>#{node.text}</strong></span>)
-      else
-        yield
-      end
     end
 
     def convert_literal(node)
